refactor: move Notifications and PageWithNotifications
into the notifications folder
