+++ src/common/features/notifications/components/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Notification";
+export * from "./PageWithNotifications";
