+++ src/common/features/notifications/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./createApiResponse";
 export * from "./NotificationsContext";
+export { PageWithNotifications } from "./components";
