+++ src/components/index.tsx
@@ -1,7 +1,5 @@
 export * from "./Layout";
 export * from "./Navigation";
-export * from "./Notification";
-export * from "./PageWithNotifications";
 export * from "./TextField";
 export * from "./TopBar";
 export * from "./Select";
